spacecraft-anomaly-detection/requirements.txt
# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Data Processing
pyarrow>=12.0.0
h5py>=3.8.0

# Configuration & Utils
pyyaml>=6.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Jupyter Support
jupyter>=1.0.0
ipywidgets>=8.0.0

# HTTP requests for data download
requests>=2.31.0

# Code Quality
black>=23.0.0
isort>=5.12.0
pytest>=7.4.0

# Optional: Weights & Biases for experiment tracking
# wandb>=0.15.0
